move sequentialFetcher to a dedicated file
